[FEAT] 1. Stop from auto archiving if batch contains same assigned pen for 2 different pen requests. 2. Stop user from archiving batches when there is a batch which contains same assigned pen for 2 different pen requests.
