+++ api/src/test/resources/application.properties
@@ -4,7 +4,7 @@ logging.level.root=ERROR
 logging.level.org.hibernate=ERROR
 logging.level.ca.bc.gov.educ.penreg=INFO
 #spring.jpa.properties.hibernate.generate_statistics=false
-spring.jpa.show-sql=true
+spring.jpa.show-sql=false
 # SQL statements and parameters
 #logging.level.org.hibernate.type.descriptor.sql=trace
 spring.main.allow-bean-definition-overriding=true
